specgen: ensure operation ids are unique
At the moment, the operation id is constructed as a remote method name.
However, it is posible to map a single method to multiple HTTP
endpoints (swagger paths)

For example, `PersistedModel.exists` is mapped to both

    HEAD /mymodels/{id}
    GET /mymodels/{id}/exists

As a result, the generated swagger operation ids are not unique.

This commit fixes the problem by detecting multi-endpoint methods
and appending HTTP verb and path to the name-based operation id.

Using the example above, the following operation ids are generated:

    Product.exists__head_Products_{id}
    Product.exists__get_Products_{id}_exists